+++ src/connected/Viewer.tsx
@@ -29,13 +29,14 @@ import {AppState} from '../states/appState';
 import {
     baseMapLayerSelector,
     imageSmoothingSelector,
-    mapProjectionSelector, selectedDatasetBoundaryLayerSelector,
+    mapProjectionSelector,
+    selectedDatasetBoundaryLayerSelector,
     selectedDatasetPlaceGroupLayersSelector,
     selectedDatasetRgbLayerSelector,
     selectedDatasetVariableLayerSelector,
     selectedPlaceGroupPlacesSelector
 } from '../selectors/controlSelectors';
-import {addUserPlace} from '../actions/dataActions';
+import { addUserPlace, addUserPlacesFromText } from '../actions/dataActions';
 import Viewer from '../components/Viewer';
 import {userPlaceGroupSelector} from "../selectors/dataSelectors";
 import {selectPlace} from "../actions/controlActions";
@@ -63,6 +64,7 @@ const mapStateToProps = (state: AppState) => {
 
 const mapDispatchToProps = {
     addUserPlace,
+    addUserPlacesFromText,
     selectPlace,
 };